Remove requirement for argparse_addons.

+++ cantools/subparsers/decode.py
@@ -2,7 +2,6 @@ import sys
 import re
 import binascii
 import struct
-from argparse_addons import Integer
 
 from .. import database
 from .utils import format_message_by_frame_id
@@ -91,7 +90,7 @@ def add_subparser(subparsers):
         help='Skip database consistency checks.')
     decode_parser.add_argument(
         '-m', '--frame-id-mask',
-        type=Integer(0),
+        type=int,
         help=('Only compare selected frame id bits to find the message in the '
               'database. By default the candump and database frame ids must '
               'be equal for a match.'))
