+++ cantools/subparsers/plot.py
@@ -47,7 +47,6 @@ import binascii
 import struct
 import datetime
 import argparse
-from argparse_addons import Integer
 try:
     from matplotlib import pyplot as plt
 except ImportError:
@@ -557,7 +556,7 @@ def add_subparser(subparsers):
         help='Skip database consistency checks.')
     decode_parser.add_argument(
         '-m', '--frame-id-mask',
-        type=Integer(0),
+        type=int,
         help=('Only compare selected frame id bits to find the message in the '
               'database. By default the candump and database frame ids must '
               'be equal for a match.'))
